Next-Gen Search Engine/pyproject.toml
[project]
name = "brightdata-video-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
